refactor: all project files
