+++ index.js
@@ -1,22 +1,1 @@
-import React, { Component } from 'react';
-
-import ReactDOM from 'react-dom';
-
-import ShopContainer from './src/ShopContainer';
-
-class App extends Component {
-  constructor(props) {
-    super(props);
-  }
-
-  render() {
-    return (
-      <ShopContainer />
-    );
-  }
-}
-
-ReactDOM.render(
-  <App />,
-  document.getElementById('root')
-);
+import './src/index.js';
