+++ package.json
@@ -99,6 +99,7 @@
     "semver": "^5.7.1",
     "source-map": "^0.6.1",
     "source-map-support": "^0.5.21",
+    "src": "^1.1.2",
     "stable": "^0.1.8",
     "supports-color": "^7.2.0",
     "svgo": "^2.8.0",
